Install firedrake-rtree, which will replace rtree on the next firedrake release

+++ firedrake/build.sh
@@ -45,12 +45,17 @@ make -j $(nproc) install
 cd && rm -rf /tmp/libspatialindex-src
 
 # rtree
-git clone https://github.com/Toblerity/rtree.git /tmp/rtree-src
-cd /tmp/rtree-src
-sed -i "s|INSTALL_PREFIX_IN|${INSTALL_PREFIX}|g" $REPODIR/firedrake/patches/08-install-prefix-in-rtree
-patch -p 1 < $REPODIR/firedrake/patches/08-install-prefix-in-rtree
-PYTHONUSERBASE=$INSTALL_PREFIX python3 -m pip install --user .
-cd && rm -rf /tmp/rtree-src
+if [[ "$RELEASE_TYPE" == "release" ]]; then
+    # firedrake dropped dependency on rtree after releasing 2026.4.0.
+    # Drop rtree after 2026.10.0 will be released, and also remove the
+    # corresponding patch 08-install-prefix-in-rtree
+    git clone https://github.com/Toblerity/rtree.git /tmp/rtree-src
+    cd /tmp/rtree-src
+    sed -i "s|INSTALL_PREFIX_IN|${INSTALL_PREFIX}|g" $REPODIR/firedrake/patches/08-install-prefix-in-rtree
+    patch -p 1 < $REPODIR/firedrake/patches/08-install-prefix-in-rtree
+    PYTHONUSERBASE=$INSTALL_PREFIX python3 -m pip install --user .
+    cd && rm -rf /tmp/rtree-src
+fi
 
 # libsupermesh
 git clone https://github.com/firedrakeproject/libsupermesh.git /tmp/libsupermesh-src
@@ -63,11 +68,27 @@ fi
 PYTHONUSERBASE=$INSTALL_PREFIX python3 -m pip install --check-build-dependencies --no-build-isolation --user .
 cd && rm -rf /tmp/libsupermesh-src
 
-# firedrake build dependencies
-PYTHONUSERBASE=$INSTALL_PREFIX python3 -m pip install --user hatchling meson meson-python pkgconfig "petsctools @ git+https://github.com/firedrakeproject/petsctools.git"
+# petsctools
+git clone https://github.com/firedrakeproject/petsctools.git /tmp/petsctools-src
+cd /tmp/petsctools-src
+if [[ "$RELEASE_TYPE" == "release" ]]; then
+    git checkout 2026.0
+else
+    git checkout main
+fi
+PYTHONUSERBASE=$INSTALL_PREFIX python3 -m pip install --check-build-dependencies --no-build-isolation --user .
+cd && rm -rf /tmp/petsctools-src
 
-# firedrake requires setuptools >= 77 to support PEP 639
-PYTHONUSERBASE=$INSTALL_PREFIX python3 -m pip install --user "setuptools>=77"
+# firedrake-rtree
+git clone https://github.com/firedrakeproject/firedrake-rtree.git /tmp/firedrake-rtree-src
+cd /tmp/firedrake-rtree-src
+if [[ "$RELEASE_TYPE" == "release" ]]; then
+    git checkout 2026.0
+else
+    git checkout main
+fi
+PYTHONUSERBASE=$INSTALL_PREFIX python3 -m pip install --check-build-dependencies --no-build-isolation --user .
+cd && rm -rf /tmp/firedrake-rtree-src
 
 # firedrake
 git clone https://github.com/firedrakeproject/firedrake.git /tmp/firedrake-src
@@ -82,18 +103,15 @@ patch -p 1 < $REPODIR/firedrake/patches/04-hardcode-omp-num-threads-in-firedrake
 PYTHONUSERBASE=$INSTALL_PREFIX python3 -m pip install --check-build-dependencies --no-build-isolation --user .
 cd && rm -rf /tmp/firedrake-src
 
-# fireshape dependencies (real mode only)
-# We package them for simplicity with firedrake so that fireshape may be pip installed.
-if [[ "$SCALAR_TYPE" != "complex" ]]; then
-    # pyroltrilinos
-    git clone https://github.com/angus-g/pyrol /tmp/pyroltrilinos-src
-    cd /tmp/pyroltrilinos-src
-    patch -p 1 < $REPODIR/firedrake/patches/06-use-system-pybind11-in-pyrol
-    patch -p 1 < $REPODIR/firedrake/patches/07-drop-setuptools_scm-in-pyrol
-    export PYBIND11_DIR=$INSTALL_PREFIX
-    PYTHONUSERBASE=$INSTALL_PREFIX CXX="mpicxx" python3 -m pip install --check-build-dependencies --no-build-isolation --user .
-    cd && rm -rf /tmp/pyroltrilinos-src
-fi
+# fireshape and icepack depend on pyroltrilinos
+# We package it for simplicity with firedrake so that fireshape and icepack may be pip installed.
+git clone https://github.com/angus-g/pyrol /tmp/pyroltrilinos-src
+cd /tmp/pyroltrilinos-src
+patch -p 1 < $REPODIR/firedrake/patches/06-use-system-pybind11-in-pyrol
+patch -p 1 < $REPODIR/firedrake/patches/07-drop-setuptools_scm-in-pyrol
+export PYBIND11_DIR=$INSTALL_PREFIX
+PYTHONUSERBASE=$INSTALL_PREFIX CXX="mpicxx" python3 -m pip install --check-build-dependencies --no-build-isolation --user .
+cd && rm -rf /tmp/pyroltrilinos-src
 
 # gusto and icepack depend on netCDF4-python
 # We package it for simplicity with firedrake so that gusto and icepack may be pip installed.
